Pretty stuff ^^

+++ scripts/emerge.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-SERVER_URL="https://geminio.rnl.tecnico.ulisboa.pt"
+SERVER_URL="https://geminio.rnl.tecnico.ulisboa.pt/packages"
 DOWNLOAD_DIR="/root"
 
 YELLOW="\e[33m"
@@ -15,16 +15,31 @@ warning() {
 	echo -e $RED"$@"$NORMAL
 }
 
+list_packages() {
+	echo -e "${YELLOW}Packages available on ${SERVER_URL}/:${NORMAL}"
+	/bin/wget --quiet "${SERVER_URL}/" -O - | awk -F "\"" '/^<a href/{sub(".tar.gz", ""); print " - " $2}'
+}
+
 pkg="${1}"
 file="${pkg}.tar.gz"
 
+if [ -z "${pkg}" -o "${pkg}" = "--list" -o "${pkg}" = "--help" ]; then
+	echo "Usage: emerge <package name>"
+	echo "Example: emerge htop"
+	list_packages
+	exit
+fi
+
+
 info "${pkg} - Downloading...     "
-wget_output=$(/bin/wget --no-verbose "${SERVER_URL}/packages/${file}" -O "${DOWNLOAD_DIR}/${file}" 2>&1)
+wget_output=$(/bin/wget --no-verbose "${SERVER_URL}/${file}" -O "${DOWNLOAD_DIR}/${file}" 2>&1)
 
 if [ $? -ne 0 ]; then
 	warning "Package '${pkg}' not found!"
 	rm "${DOWNLOAD_DIR}/${file}"
 	echo "$wget_output"
+	echo
+	list_packages
 	exit
 fi
